# Huge speed increase via compile optimization