--- a/sns_ik_examples/launch/test_ik_solvers.launch
+++ b/sns_ik_examples/launch/test_ik_solvers.launch
@@ -1,13 +1,13 @@
 <?xml version="1.0"?>
 <launch>
-  <arg name="num_samples_pos" default="200" />
+  <arg name="num_samples_pos" default="1000" />
   <arg name="num_samples_vel" default="1000" />
   <arg name="chain_start" default="base" />
   <arg name="chain_end" default="right_hand" />
   <arg name="timeout" default="0.005" />
   <arg name="urdf_param" default="/robot_description" />
   <arg name="use_random_position_seed" default="false"/>
-  <arg name="use_delta_position_seed" default="true"/>
+  <arg name="use_delta_position_seed" default="false"/>
   <arg name="delta_position_seed_value" default="0.2"/>
 
   <node name="sns_ik_tests" pkg="sns_ik_examples" type="all_ik_tests" output="screen">