Turned on compile optimization flag -O2 and saw a ~20X time decrease. SNS Fast is now slightly faster than TRAC-IK though with a reduced success rate starting with a nominal joint seed of the joint mid-points.

Current metrics:

************************************
Position IK Summary:  Nominal Seed
SNS: 83.10% success rate with (time: 0.97 ± 0.86 ms)
SNS Optimal Scale Margin: 83.80% success rate with (time: 1.56 ± 1.60 ms)
SNS Optimal: 83.10% success rate with (time: 1.14 ± 1.21 ms)
SNS Fast: 84.00% success rate with (time: 0.44 ± 0.37 ms)
SNS Fast Optimal: 83.80% success rate with (time: 0.68 ± 0.57 ms)
KDL: 28.20% success rate with (time: 1.33 ± 0.62 ms)
TRAC: 99.80% success rate with (time: 0.49 ± 0.45 ms)

************************************
Velocity IK Summary:
SNS: 97.00% w/o and 100.00% w/ scaling success rates with an average time of 0.024 ms
SNS Optimal Scale Margin: 41.20% w/o and 100.00% w/ scaling success rates with an average time of 0.044 ms
SNS Optimal: 97.00% w/o and 100.00% w/ scaling success rates with an average time of 0.026 ms
SNS Fast: 97.00% w/o and 100.00% w/ scaling success rates with an average time of 0.013 ms
SNS Fast Optimal: 57.90% w/o and 100.00% w/ scaling success rates with an average time of 0.019 ms
KDL Velocity: 71.70% w/o and 71.70% w/ scaling success rates with an average time of 0.018 ms

************************************
Position IK Summary:  Random Seeds
SNS: 45.30% success rate with (time: 1.80 ± 1.22 ms)
SNS Optimal Scale Margin: 45.40% success rate with (time: 2.77 ± 2.16 ms)
SNS Optimal: 45.90% success rate with (time: 2.16 ± 1.65 ms)
SNS Fast: 45.30% success rate with (time: 0.83 ± 0.46 ms)
SNS Fast Optimal: 45.50% success rate with (time: 1.25 ± 0.81 ms)
KDL: 27.30% success rate with (time: 1.41 ± 0.64 ms)
TRAC: 99.60% success rate with (time: 0.65 ± 0.75 ms)