@rethink-forrest 2016-04-01
Position IK Summary:  Near Seed (0.2 rad per joint)
SNS: 98.00% success rate with (time: 0.16 ± 0.33 ms)
SNS Optimal Scale Margin: 98.20% success rate with (time: 0.18 ± 0.28 ms)
SNS Optimal: 97.80% success rate with (time: 0.17 ± 0.37 ms)
SNS Fast: 98.00% success rate with (time: 0.10 ± 0.18 ms)
SNS Fast Optimal: 98.60% success rate with (time: 0.11 ± 0.17 ms)
KDL: 98.30% success rate with (time: 0.12 ± 0.26 ms)
TRAC: 100.00% success rate with (time: 0.17 ± 0.14 ms)